Move progressive cache update logic to a separate file and add progressive cache metrics
