+++ consensus/state_processing/src/common/mod.rs
@@ -7,7 +7,7 @@ mod slash_validator;
 
 pub mod altair;
 pub mod base;
-pub mod initialize_progressive_balances_cache;
+pub mod update_progressive_balances_cache;
 
 pub use deposit_data_tree::DepositDataTree;
 pub use get_attestation_participation::get_attestation_participation_flag_indices;
