+++ consensus/state_processing/src/common/slash_validator.rs
@@ -1,3 +1,4 @@
+use crate::common::update_progressive_balances_cache::update_progressive_balances_on_slashing;
 use crate::{
     common::{decrease_balance, increase_balance, initiate_validator_exit},
     per_block_processing::errors::BlockProcessingError,
@@ -42,6 +43,8 @@ pub fn slash_validator<T: EthSpec>(
         validator_effective_balance
             .safe_div(spec.min_slashing_penalty_quotient_for_state(state))?,
     )?;
+
+    update_progressive_balances_on_slashing(state, slashed_index)?;
 
     // Apply proposer and whistleblower rewards
     let proposer_index = ctxt.get_proposer_index(state, spec)? as usize;
