+++ consensus/state_processing/src/metrics.rs
@@ -23,4 +23,15 @@ lazy_static! {
         "beacon_participation_prev_epoch_active_gwei_total",
         "Total effective balance (gwei) of validators active in the previous epoch"
     );
+    /*
+     * Participation Metrics (progressive balances)
+     */
+    pub static ref PARTICIPATION_PREV_EPOCH_TARGET_ATTESTING_GWEI_PROGRESSIVE_TOTAL: Result<IntGauge> = try_create_int_gauge(
+        "beacon_participation_prev_epoch_target_attesting_gwei_progressive_total",
+        "Progressive total effective balance (gwei) of validators who attested to the target in the previous epoch"
+    );
+    pub static ref PARTICIPATION_CURR_EPOCH_TARGET_ATTESTING_GWEI_PROGRESSIVE_TOTAL: Result<IntGauge> = try_create_int_gauge(
+        "beacon_participation_curr_epoch_target_attesting_gwei_progressive_total",
+        "Progressive total effective balance (gwei) of validators who attested to the target in the current epoch"
+    );
 }
