+++ consensus/state_processing/src/per_epoch_processing/altair.rs
@@ -1,5 +1,7 @@
 use super::{process_registry_updates, process_slashings, EpochProcessingSummary, Error};
-use crate::common::initialize_progressive_balances_cache::initialize_progressive_balances_cache;
+use crate::common::update_progressive_balances_cache::{
+    initialize_progressive_balances_cache, update_progressive_balances_on_epoch_transition,
+};
 use crate::per_epoch_processing::{
     effective_balance_updates::process_effective_balance_updates,
     historical_roots_update::process_historical_roots_update,
@@ -77,10 +79,7 @@ pub fn process_epoch<T: EthSpec>(
     // Rotate the epoch caches to suit the epoch transition.
     state.advance_caches(spec)?;
 
-    // Update progressive total balances
-    state
-        .progressive_balances_cache_mut()
-        .on_epoch_transition(spec)?;
+    update_progressive_balances_on_epoch_transition(state, spec)?;
 
     Ok(EpochProcessingSummary::Altair {
         participation_cache,
